Refactor Pipboy class cont. - Extract operating on pipboy into a handler
