+++ src/Pipboy.cpp
@@ -8,79 +8,10 @@
 #include "HandPose.h"
 #include "utils.h"
 #include "common/Logger.h"
-#include "f4vr/F4VRUtils.h"
-#include "f4vr/scaleformUtils.h"
 #include "f4vr/VRControllersManager.h"
 
-using namespace RE::Scaleform;
 using namespace std::chrono;
 using namespace common;
-
-namespace
-{
-    bool isPrimaryTriggerPressed()
-    {
-        return f4vr::VRControllers.isPressed(vr::k_EButton_SteamVR_Trigger, f4vr::Hand::Primary);
-    }
-
-    bool isAButtonPressed()
-    {
-        return f4vr::VRControllers.isPressed(vr::k_EButton_A, f4vr::Hand::Primary);
-    }
-
-    bool isBButtonPressed()
-    {
-        return f4vr::VRControllers.isPressed(vr::k_EButton_ApplicationMenu, f4vr::Hand::Primary);
-    }
-
-    bool isPrimaryGripPressHeldDown()
-    {
-        return f4vr::VRControllers.isPressHeldDown(vr::k_EButton_Grip, f4vr::Hand::Primary);
-    }
-
-    bool isWorldMapVisible(const GFx::Movie* root)
-    {
-        return f4vr::isElementVisible(root, "root.Menu_mc.CurrentPage.WorldMapHolder_mc");
-    }
-
-    std::string getCurrentMapPath(const GFx::Movie* root, const std::string& suffix = "")
-    {
-        return (isWorldMapVisible(root) ? "root.Menu_mc.CurrentPage.WorldMapHolder_mc" : "root.Menu_mc.CurrentPage.LocalMapHolder_mc") + suffix;
-    }
-
-    /**
-     * Is context menu message box popup is visible or not. Only one can be visible at a time.
-     */
-    bool isMessageHolderVisible(const GFx::Movie* root)
-    {
-        return root &&
-        (f4vr::isElementVisible(root, "root.Menu_mc.CurrentPage.MessageHolder_mc")
-            || f4vr::isElementVisible(root, "root.Menu_mc.CurrentPage.QuestsTab_mc.MessageHolder_mc"));
-    }
-
-    bool isQuestTabVisibleOnDataPage(const GFx::Movie* root)
-    {
-        return f4vr::isElementVisible(root, "root.Menu_mc.CurrentPage.QuestsTab_mc");
-    }
-
-    bool isQuestTabObjectiveListEnabledOnDataPage(const GFx::Movie* root)
-    {
-        GFx::Value var;
-        return root->GetVariable(&var, "root.Menu_mc.CurrentPage.QuestsTab_mc.ObjectivesList_mc.selectedIndex") && var.GetType() == GFx::Value::ValueType::kInt && var.GetInt() > -
-            1;
-    }
-
-    bool isWorkshopsTabVisibleOnDataPage(const GFx::Movie* root)
-    {
-        return f4vr::isElementVisible(root, "root.Menu_mc.CurrentPage.WorkshopsTab_mc");
-    }
-
-    GFx::Movie* getPipboyMenuRootMovie()
-    {
-        const auto menu = RE::UI::GetSingleton()->GetMenu("PipboyMenu");
-        return menu ? menu->uiMovie.get() : nullptr;
-    }
-}
 
 namespace frik
 {
@@ -153,20 +84,15 @@ namespace frik
 
         storeLastPipboyPage();
 
-        const auto pipboyRoot = getPipboyMenuRootMovie();
-        _physicalHandler.operate(_lastPipboyPage, isMessageHolderVisible(pipboyRoot));
+        _physicalHandler.operate(_lastPipboyPage);
 
         if (f4vr::isInPowerArmor()) {
             lastRadioFreq = 0.0; // Ensures radio needle doesn't get messed up when entering and then exiting Power Armor.
             return;
         }
 
-        if (_isOnStatus && pipboyRoot) {
-            if (g_config.enablePrimaryControllerPipboyUse && !g_frik.isPipboyConfigurationModeActive()) {
-                // Use primary hand controller to operate Pipboy
-                handlePrimaryControllerThumbstickOperation(pipboyRoot);
-                handlePrimaryControllerButtonsOperation(pipboyRoot, false);
-            }
+        if (_isOnStatus) {
+            _operationHandler.operate();
 
             dampenPipboyScreen();
         }
@@ -182,42 +108,6 @@ namespace frik
         // sets 3rd Person Pipboy Scale
         if (const auto pipboy3Rd = f4vr::findNode(f4vr::getWorldRootNode(), "PipboyBone")) {
             pipboy3Rd->local.scale = g_config.pipBoyScale;
-        }
-    }
-
-    /**
-     * Execute the given operation on the current Pipboy page.
-     */
-    void Pipboy::execOperation(const PipboyOperation operation)
-    {
-        const auto root = getPipboyMenuRootMovie();
-        if (!root) {
-            logger::warn<>("Pipboy operation failed, root is null");
-            return;
-        }
-
-        switch (operation) {
-        case PipboyOperation::GOTO_PREV_PAGE:
-            gotoPrevPage(root);
-            break;
-        case PipboyOperation::GOTO_NEXT_PAGE:
-            gotoNextPage(root);
-            break;
-        case PipboyOperation::GOTO_PREV_TAB:
-            gotoPrevTab(root);
-            break;
-        case PipboyOperation::GOTO_NEXT_TAB:
-            gotoNextTab(root);
-            break;
-        case PipboyOperation::MOVE_LIST_SELECTION_UP:
-            moveListSelectionUpDown(root, true);
-            break;
-        case PipboyOperation::MOVE_LIST_SELECTION_DOWN:
-            moveListSelectionUpDown(root, false);
-            break;
-        case PipboyOperation::PRIMARY_PRESS:
-            handlePrimaryControllerButtonsOperation(root, true);
-            break;
         }
     }
 
@@ -477,257 +367,14 @@ namespace frik
     }
 
     /**
-     * Execute operation on the currently active Pipboy UI by primary thumbstick input.
-     * On map page the map is moved.
-     * On other pages the list selection is moved up/down or tabs are changed left/right.
-     */
-    void Pipboy::handlePrimaryControllerThumbstickOperation(GFx::Movie* const root)
-    {
-        if (!isPrimaryGripPressHeldDown()) {
-            const auto doinantHandStick = f4vr::VRControllers.getAxisValue(f4vr::Hand::Primary, 0);
-            if (_lastPipboyPage == PipboyPage::MAP && !isMessageHolderVisible(root)) {
-                // Map Tab
-                GFx::Value akArgs[2];
-                akArgs[0] = doinantHandStick.x * -1;
-                akArgs[1] = doinantHandStick.y;
-                // Move Map
-                root->Invoke("root.Menu_mc.CurrentPage.WorldMapHolder_mc.PanMap", nullptr, akArgs, 2);
-                root->Invoke("root.Menu_mc.CurrentPage.LocalMapHolder_mc.PanMap", nullptr, akArgs, 2);
-            } else {
-                if (doinantHandStick.y > 0.85) {
-                    if (!_controlSleepStickyY) {
-                        _controlSleepStickyY = true;
-                        moveListSelectionUpDown(root, true);
-                        std::thread t2(&Pipboy::rightStickYSleep, this, 155);
-                        t2.detach();
-                    }
-                }
-                if (doinantHandStick.y < -0.85) {
-                    if (!_controlSleepStickyY) {
-                        _controlSleepStickyY = true;
-                        moveListSelectionUpDown(root, false);
-                        std::thread t2(&Pipboy::rightStickYSleep, this, 155);
-                        t2.detach();
-                    }
-                }
-                if (doinantHandStick.x < -0.85) {
-                    if (!_controlSleepStickyX) {
-                        _controlSleepStickyX = true;
-                        gotoPrevTab(root);
-                        std::thread t3(&Pipboy::rightStickXSleep, this, 170);
-                        t3.detach();
-                    }
-                }
-                if (doinantHandStick.x > 0.85) {
-                    if (!_controlSleepStickyX) {
-                        _controlSleepStickyX = true;
-                        gotoNextTab(root);
-                        std::thread t3(&Pipboy::rightStickXSleep, this, 170);
-                        t3.detach();
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Execute operation on the currently active Pipboy UI by primary controller buttons.
-     * First thing is to detect is a message box is visible (context menu options) as message box and main lists are active at the
-     * same time. So, if the message box is visible we will only operate on the message box list and not the main list.
-     * See documentation: https://github.com/rollingrock/Fallout-4-VR-Body/wiki/Development-%E2%80%90-Pipboy-Controls
-     * @param root - Pipboy Scaleform UI root
-     * @param triggerPressed - true if operate as trigger is pressed regardless of controller input (used for physical handler)
-     */
-    void Pipboy::handlePrimaryControllerButtonsOperation(GFx::Movie* root, bool triggerPressed)
-    {
-        // page level navigation
-        const bool gripPressHeldDown = isPrimaryGripPressHeldDown();
-        if (!gripPressHeldDown && isAButtonPressed()) {
-            gotoPrevPage(root);
-            return;
-        }
-        if (!gripPressHeldDown && isBButtonPressed()) {
-            gotoNextPage(root);
-            return;
-        }
-
-        // include actual controller check
-        triggerPressed = triggerPressed || isPrimaryTriggerPressed();
-
-        // Context menu message box handling
-        if (triggerPressed && isMessageHolderVisible(root)) {
-            triggerShortHeptic();
-            f4vr::doOperationOnScaleformMessageHolderList(root, "root.Menu_mc.CurrentPage.MessageHolder_mc", f4vr::ScaleformListOp::Select);
-            f4vr::doOperationOnScaleformMessageHolderList(root, "root.Menu_mc.CurrentPage.QuestsTab_mc.MessageHolder_mc", f4vr::ScaleformListOp::Select);
-            // prevent affecting the main list if message box is visible
-            return;
-        }
-
-        // Specific handling by current page
-        switch (f4vr::getScaleformInt(root, "root.Menu_mc.DataObj._CurrentPage").value_or(-1)) {
-        case 0:
-            handlePrimaryControllerOperationOnStatusPage(root, triggerPressed);
-            break;
-        case 1:
-            handlePrimaryControllerOperationOnInventoryPage(root, triggerPressed);
-            break;
-        case 2:
-            handlePrimaryControllerOperationOnDataPage(root, triggerPressed);
-            break;
-        case 3:
-            handlePrimaryControllerOperationOnMapPage(root, triggerPressed);
-            break;
-        case 4:
-            handlePrimaryControllerOperationOnRadioPage(root, triggerPressed);
-            break;
-        default: ;
-        }
-
-        if (f4vr::VRControllers.isPressed(vr::EVRButtonId::k_EButton_Axis0, f4vr::Hand::Primary)) {
-            root->Invoke("root.Menu_mc.CurrentPage.onMessageButtonPress()", nullptr, nullptr, 0);
-        }
-    }
-
-    void Pipboy::gotoPrevPage(GFx::Movie* root)
-    {
-        root->Invoke("root.Menu_mc.gotoPrevPage", nullptr, nullptr, 0);
-    }
-
-    void Pipboy::gotoNextPage(GFx::Movie* root)
-    {
-        root->Invoke("root.Menu_mc.gotoNextPage", nullptr, nullptr, 0);
-    }
-
-    void Pipboy::gotoPrevTab(GFx::Movie* root)
-    {
-        triggerShortHeptic();
-        root->Invoke("root.Menu_mc.gotoPrevTab", nullptr, nullptr, 0);
-    }
-
-    void Pipboy::gotoNextTab(GFx::Movie* root)
-    {
-        triggerShortHeptic();
-        root->Invoke("root.Menu_mc.gotoNextTab", nullptr, nullptr, 0);
-    }
-
-    /**
-     * Execute moving selection on the currently active list up or down.
-     * Whatever the list found to exist it will be moved, mostly we can "try" to move any of the lists and only the one that exists will be moved.
-     * On DATA page all 3 tabs can exist at the same time, so we need to check which one is visible to prevent working on a hidden one.
-     * First thing is to detect is a message box is visible (context menu options) as message box and main lists are active at the
-     * same time. So, if the message box is visible we will only operate on the message box list and not the main list.
-     */
-    void Pipboy::moveListSelectionUpDown(GFx::Movie* root, const bool moveUp)
-    {
-        triggerShortHeptic();
-
-        const auto listOp = moveUp ? f4vr::ScaleformListOp::MoveUp : f4vr::ScaleformListOp::MoveDown;
-
-        if (isMessageHolderVisible(root)) {
-            f4vr::doOperationOnScaleformMessageHolderList(root, "root.Menu_mc.CurrentPage.MessageHolder_mc", listOp);
-            f4vr::doOperationOnScaleformMessageHolderList(root, "root.Menu_mc.CurrentPage.QuestsTab_mc.MessageHolder_mc", listOp);
-            // prevent affecting the main list if message box is visible
-            return;
-        }
-
-        // Inventory, Radio, Special, and Perks tabs
-        f4vr::doOperationOnScaleformList(root, "root.Menu_mc.CurrentPage.List_mc", listOp);
-        f4vr::doOperationOnScaleformList(root, "root.Menu_mc.CurrentPage.SPECIALTab_mc.List_mc", listOp);
-        f4vr::doOperationOnScaleformList(root, "root.Menu_mc.CurrentPage.PerksTab_mc.List_mc", listOp);
-
-        // Quest, Workshop, and Stats tabs exist at the same time, need to check which one is visible
-        if (isQuestTabVisibleOnDataPage(root)) {
-            // Quests tab has 2 lists for the main quests and quest objectives
-            const char* listPath = isQuestTabObjectiveListEnabledOnDataPage(root)
-                ? "root.Menu_mc.CurrentPage.QuestsTab_mc.ObjectivesList_mc"
-                : "root.Menu_mc.CurrentPage.QuestsTab_mc.QuestsList_mc";
-            f4vr::doOperationOnScaleformList(root, listPath, listOp);
-        } else if (isWorkshopsTabVisibleOnDataPage(root)) {
-            f4vr::doOperationOnScaleformList(root, "root.Menu_mc.CurrentPage.WorkshopsTab_mc.List_mc", listOp);
-        } else {
-            f4vr::doOperationOnScaleformList(root, "root.Menu_mc.CurrentPage.StatsTab_mc.CategoryList_mc", listOp);
-        }
-    }
-
-    void Pipboy::handlePrimaryControllerOperationOnStatusPage(GFx::Movie* root, const bool triggerPressed)
-    {
-        if (triggerPressed) {
-            triggerShortHeptic();
-            f4vr::doOperationOnScaleformList(root, "root.Menu_mc.CurrentPage.SPECIALTab_mc.List_mc", f4vr::ScaleformListOp::Select);
-            f4vr::doOperationOnScaleformList(root, "root.Menu_mc.CurrentPage.PerksTab_mc.List_mc", f4vr::ScaleformListOp::Select);
-        }
-    }
-
-    void Pipboy::handlePrimaryControllerOperationOnInventoryPage(GFx::Movie* root, const bool triggerPressed)
-    {
-        if (triggerPressed) {
-            triggerShortHeptic();
-            f4vr::doOperationOnScaleformList(root, "root.Menu_mc.CurrentPage.List_mc", f4vr::ScaleformListOp::Select);
-        }
-    }
-
-    /**
-     * On DATA page all 3 tabs can exist at the same time, so we need to check which one is visible to prevent working on a hidden one.
-     */
-    void Pipboy::handlePrimaryControllerOperationOnDataPage(GFx::Movie* root, const bool triggerPressed)
-    {
-        if (triggerPressed) {
-            triggerShortHeptic();
-            if (isQuestTabVisibleOnDataPage(root)) {
-                f4vr::doOperationOnScaleformList(root, "root.Menu_mc.CurrentPage.QuestsTab_mc.QuestsList_mc", f4vr::ScaleformListOp::Select);
-            } else {
-                // open quest in map
-                f4vr::invokeScaleformProcessUserEvent(root, "root.Menu_mc.CurrentPage.WorkshopsTab_mc", "XButton");
-            }
-        }
-    }
-
-    /**
-     * For handling map faster travel or marker setting we need to know if fast travel can be done using "bCanFastTravel"
-     * and then let the Pipboy code handle the rest by sending the right event to the currently visible map (world/local).
-     */
-    void Pipboy::handlePrimaryControllerOperationOnMapPage(GFx::Movie* root, const bool triggerPressed)
-    {
-        if (isPrimaryGripPressHeldDown()) {
-            if (triggerPressed) {
-                // switch world/local maps
-                triggerShortHeptic();
-                f4vr::invokeScaleformProcessUserEvent(root, "root.Menu_mc.CurrentPage", "XButton");
-            } else {
-                // zoom map
-                const auto [_, primAxisY] = f4vr::VRControllers.getAxisValue(f4vr::Hand::Primary);
-                if (fNotEqual(primAxisY, 0, 0.5f)) {
-                    GFx::Value args[1];
-                    args[0] = primAxisY / 100.f;
-                    root->Invoke(getCurrentMapPath(root, ".ZoomMap").c_str(), nullptr, args, 1);
-                }
-            }
-        } else if (triggerPressed) {
-            triggerShortHeptic();
-
-            // handle fast travel, custom marker
-            const char* eventName = f4vr::getScaleformBool(root, getCurrentMapPath(root, ".bCanFastTravel").c_str()) ? "MapHolder:activate_marker" : "MapHolder:set_custom_marker";
-            f4vr::invokeScaleformDispatchEvent(root, getCurrentMapPath(root), eventName);
-        }
-    }
-
-    void Pipboy::handlePrimaryControllerOperationOnRadioPage(GFx::Movie* root, const bool triggerPressed)
-    {
-        if (triggerPressed) {
-            triggerShortHeptic();
-            f4vr::doOperationOnScaleformList(root, "root.Menu_mc.CurrentPage.List_mc", f4vr::ScaleformListOp::Select);
-        }
-    }
-
-    /**
      * Get Current Pipboy Tab and store it.
+     * For handling if Pipboy was opened in PA where most of this code is not executed.
      */
     void Pipboy::storeLastPipboyPage()
     {
-        const auto root = getPipboyMenuRootMovie();
-        GFx::Value PBCurrentPage;
-        if (root && root->GetVariable(&PBCurrentPage, "root.Menu_mc.DataObj._CurrentPage") && PBCurrentPage.GetType() != GFx::Value::ValueType::kUndefined) {
-            _lastPipboyPage = static_cast<PipboyPage>(PBCurrentPage.GetUInt());
+        const auto pipboyPage = PipboyOperationHandler::getCurrentPipboyPage(PipboyOperationHandler::getPipboyMenuRoot());
+        if (pipboyPage.has_value()) {
+            _lastPipboyPage = pipboyPage.value();
         }
     }
 
@@ -760,24 +407,6 @@ namespace frik
             _pipboyScreenPrevFrame = pipboyScreen->world;
             f4vr::updateDown(pipboyScreen, false);
         }
-    }
-
-    /**
-     * Prevents continuous Input from Right Stick X Axis
-     */
-    void Pipboy::rightStickXSleep(const int time)
-    {
-        Sleep(time);
-        _controlSleepStickyX = false;
-    }
-
-    /**
-     * Prevents continuous Input from Right Stick Y Axis
-     */
-    void Pipboy::rightStickYSleep(const int time)
-    {
-        Sleep(time);
-        _controlSleepStickyY = false;
     }
 
     /**
